Progress.

+++ classes/task/cron_task_gradesync.php
@@ -64,10 +64,9 @@ class cron_task_gradesync extends \core\task\scheduled_task {
         }
         $this->mappingcourses = array_unique(array_column($mappings, 'courseid'));
 
+        $this->cleanup_mappings();
 
         $this->sync_grades();
-
-        $this->cleanup_mappings();
         
         $this->log_finish("Done");
 
@@ -101,34 +100,37 @@ class cron_task_gradesync extends \core\task\scheduled_task {
 
     /**
      * Delete mappings where course/group no longer available.
-     *
      */
     protected function cleanup_mappings() {
         global $DB;
 
+
+        $this->log("Starting mapping cleanup", 1);
         // Get list of visible and active courseids.
         $now = time();
         $sql = "SELECT id
                   FROM {course}
                  WHERE visible = 1
                    AND (enddate = 0 OR enddate > {$now})";
-        $courseids = array_values($DB->get_records_sql($sql));
+        $courseids = array_values(array_column($DB->get_records_sql($sql), 'id'));
         // Determine inactive courses.
         $activemappings = array_intersect($courseids, $this->mappingcourses);
-        $inactivemappings = array_diff($activemappings, $this->mappingcourses);
+        $inactivemappings = array_diff($this->mappingcourses, $activemappings);
         // Delete inactive mappings.
-        $this->log("Delete mappings for inactive courses: " . implode(', ', $inactivemappings), 2);
-        list($insql, $inparams) = $DB->get_in_or_equal($inactivemappings);
-        $sql = "DELETE FROM {gradesync_mappings} WHERE courseid $insql";
-        $DB->execute($sql, $inparams);
+        if ($inactivemappings) {
+            $this->log("Delete mappings for inactive courses: " . implode(', ', $inactivemappings), 2);
+            list($insql, $inparams) = $DB->get_in_or_equal($inactivemappings);
+            $sql = "DELETE FROM {gradesync_mappings} WHERE courseid $insql";
+            $DB->execute($sql, $inparams);
+        }
 
-
-        // Get list of viable course groups.
+        // Get list of viable groups.
         list($insql, $inparams) = $DB->get_in_or_equal($courseids);
         $sql = "SELECT id
                   FROM {groups}
                  WHERE courseid $insql";
-        $groupids = array_values($DB->get_records_sql($sql));
+        $groupids = array_values(array_column($DB->get_records_sql($sql, $inparams), 'id'));
+
         // Determine inactive groups.
         $sql = "SELECT id, groupid
                   FROM {gradesync_mappings}";
@@ -138,12 +140,16 @@ class cron_task_gradesync extends \core\task\scheduled_task {
         }
         $mappinggroups = array_unique(array_column($mappings, 'groupid'));
         $activemappings = array_intersect($groupids, $mappinggroups);
-        $inactivemappings = array_diff($activemappings, $mappinggroups);
+        $activemappings[] = 0; // Course level mappings have always default to 0 for groupid.
+        $inactivemappings = array_diff($mappinggroups, $activemappings);
         // Delete inactive mappings.
-        $this->log("Delete mappings for inactive groups: " . implode(', ', $inactivemappings), 2);
-        list($insql, $inparams) = $DB->get_in_or_equal($inactivemappings);
-        $sql = "DELETE FROM {gradesync_mappings} WHERE groupid $insql";
-        $DB->execute($sql, $inparams);
+        if ($inactivemappings) {
+            $this->log("Delete mappings for inactive groups: " . implode(', ', $inactivemappings), 2);
+            list($insql, $inparams) = $DB->get_in_or_equal($inactivemappings);
+            $sql = "DELETE FROM {gradesync_mappings} WHERE groupid $insql";
+            $DB->execute($sql, $inparams);
+        }
+        
 
     }
